Fix #1, update to bplib development branch (BPv7)
Batch commit of updates for initial compatibility with development
branch of BPLIB, which supports BPv7.  Currently a single CLA is
hardcoded, only the application data flows are specified in the flow
table.

This should work in conjunction with the CF app and two instance of CFS
to demonstrate a file transfer across nodes..