Fix #21, add build and run workflow
A complete functional workflow that builds two instances of CFS
with CF, BP, and BPLib, including configuration tables that are set
up for bundle transfer between the two instances.

A small test file is generated on the first instance, and its md5sum is
computed.  Then both instances of CFS are executed, the flows are
enabled, and a CFDP transfer is initiated to send the file.  Then both
CFS instances are shut down, and the md5sum of the test file is checked
on the recieve side, and it should match.